Started with drawing a rough schematic for a hub
Started drawing a schematic of how a USB3 hub with support for toggling
the ports via USB-Command and Buttons could look like.
The Idea is, that the HUB will toggle power output, but can be written
via SMBUS from the RP2040, which is also connected via a port to the
USB-Hub, to give feedback about power drawn from the other ports to the
system for current consumption montoring.
Only observing 3 Ports for now to see if it is viable, and also kind of
because the PI only has 4 ADC channels, but I wanted to have symetrie.